CPM-558: Move Association Intents to a subfolder
