+++ src/Akeneo/Pim/Enrichment/Product/back/API/Command/UserIntent/Association/AssociationUserIntent.php
@@ -2,13 +2,15 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Pim\Enrichment\Product\API\Command\UserIntent;
+namespace Akeneo\Pim\Enrichment\Product\API\Command\UserIntent\Association;
+
+use Akeneo\Pim\Enrichment\Product\API\Command\UserIntent\UserIntent;
 
 /**
  * @copyright 2022 Akeneo SAS (https://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-interface AssociationsUserIntent extends UserIntent
+interface AssociationUserIntent extends UserIntent
 {
     public function associationType(): string;
 }
